refactor: extract publishing image form section
