+++ tests/unit/test_init.py
@@ -3281,7 +3281,7 @@ class TestWantedShippingSelection:
             patch.object(test_bot, "_set_special_attributes", new_callable = AsyncMock),
             patch.object(test_bot, "_set_shipping", new_callable = AsyncMock),
             patch("kleinanzeigen_bot.publishing_form.set_contact_fields", new_callable = AsyncMock),
-            patch.object(test_bot, "_upload_images", new_callable = AsyncMock),
+            patch("kleinanzeigen_bot.publishing_form.fill_image_section", new_callable = AsyncMock),
             patch.object(test_bot, "web_set_input_value", new_callable = AsyncMock),
             patch("kleinanzeigen_bot.captcha_flow.check_and_wait_for_captcha", new_callable = AsyncMock),
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None),
@@ -3578,379 +3578,3 @@ class TestBuyNowRadioWarning:
             self._mock_publish_ad_dependencies(test_bot, mock_page, probe_side_effect, check_side_effect),
         ):
             await test_bot.publish_ad(ad_file, ad_cfg, ad_cfg_orig, [], AdUpdateStrategy.REPLACE)
-
-
-class TestImageUploadProcessedMarkerFallback:
-    """Regression tests for image upload completion detection via hidden marker inputs."""
-
-    @staticmethod
-    def _build_two_image_ad(base_ad_config:dict[str, Any], tmp_path:Path) -> tuple[Ad, str, str]:
-        image_a = tmp_path / "img_a.jpg"
-        image_b = tmp_path / "img_b.jpg"
-        image_a.write_bytes(b"")
-        image_b.write_bytes(b"")
-        ad_cfg = Ad.model_validate(base_ad_config | {"images": [str(image_a), str(image_b)]})
-        return ad_cfg, str(image_a), str(image_b)
-
-    @staticmethod
-    def _build_marker(url:str) -> MagicMock:
-        marker = MagicMock()
-        marker.attrs.value = url
-        return marker
-
-    @staticmethod
-    @contextmanager
-    def _mock_upload_dependencies(
-        test_bot:KleinanzeigenBot,
-        file_input:MagicMock,
-        find_all_side_effect:Callable[..., Awaitable[list[MagicMock]]],
-        await_side_effect:Callable[..., Awaitable[Any]],
-    ) -> Iterator[None]:
-        async def find_all_once_side_effect(selector_type:By, selector_value:str, *_:Any, **__:Any) -> list[MagicMock]:
-            return await find_all_side_effect(selector_type, selector_value, **__)
-
-        with (
-            patch.object(test_bot, "web_find", new_callable = AsyncMock, return_value = file_input),
-            patch.object(test_bot, "_web_find_all_once", new_callable = AsyncMock, side_effect = find_all_once_side_effect),
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_await", new_callable = AsyncMock, side_effect = await_side_effect),
-        ):
-            yield
-
-    @pytest.mark.asyncio
-    async def test_upload_images_succeeds_with_hidden_markers_when_thumbnails_absent(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-    ) -> None:
-        """Hidden adImages markers should satisfy completion when thumbnail list is missing."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        file_input = MagicMock()
-        file_input.send_file = AsyncMock()
-
-        marker_a = self._build_marker("https://img.example/a.jpg")
-        marker_b = self._build_marker("https://img.example/b.jpg")
-        marker_query_count = 0
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, *_:Any, **__:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    return []  # baseline before upload
-                return [marker_a, marker_b]
-            return []
-
-        async def await_side_effect(condition:Callable[[], Awaitable[bool]], **_:Any) -> bool:
-            if await condition():
-                return True
-            raise TimeoutError("condition did not pass")
-
-        with self._mock_upload_dependencies(test_bot, file_input, find_all_side_effect, await_side_effect):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        file_input.send_file.assert_any_await(image_a)
-        file_input.send_file.assert_any_await(image_b)
-
-    @pytest.mark.asyncio
-    async def test_upload_images_refetches_file_input_per_image_to_avoid_stale_element(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-    ) -> None:
-        """Each image upload should re-fetch the file input because the DOM replaces it after selection."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        first_file_input = MagicMock()
-        first_file_input.send_file = AsyncMock()
-        second_file_input = MagicMock()
-        second_file_input.send_file = AsyncMock()
-
-        marker_a = self._build_marker("https://img.example/a.jpg")
-        marker_b = self._build_marker("https://img.example/b.jpg")
-        marker_query_count = 0
-
-        async def find_side_effect(selector_type:By, selector_value:str, **_:Any) -> MagicMock:
-            assert selector_type == By.CSS_SELECTOR
-            assert selector_value == "input[type=file]"
-            if first_file_input.send_file.await_count == 0:
-                return first_file_input
-            return second_file_input
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, *_:Any, **__:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    return []  # baseline before upload
-                return [marker_a, marker_b]
-            return []
-
-        async def await_side_effect(condition:Callable[[], Awaitable[bool]], **_:Any) -> bool:
-            if await condition():
-                return True
-            raise TimeoutError("condition did not pass")
-
-        with (
-            patch.object(test_bot, "web_find", new_callable = AsyncMock, side_effect = find_side_effect) as mock_find,
-            patch.object(test_bot, "_web_find_all_once", new_callable = AsyncMock, side_effect = find_all_side_effect),
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_await", new_callable = AsyncMock, side_effect = await_side_effect),
-        ):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        first_file_input.send_file.assert_awaited_once_with(image_a)
-        second_file_input.send_file.assert_awaited_once_with(image_b)
-        assert mock_find.await_count >= 2
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        ("baseline_count", "post_count", "expected_found"),
-        [
-            pytest.param(2, 2, 0, id = "stale-only-markers"),
-            pytest.param(0, 1, 1, id = "one-new-marker"),
-        ],
-    )
-    async def test_upload_images_timeout_reports_processed_count(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-        baseline_count:int,
-        post_count:int,
-        expected_found:int,
-    ) -> None:
-        """Upload timeout should report the correct processed-marker count based on baseline vs post-upload markers."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        file_input = MagicMock()
-        file_input.send_file = AsyncMock()
-
-        marker_query_count = 0
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, **_:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    return [self._build_marker(f"https://img.example/baseline-{i}.jpg") for i in range(baseline_count)]
-                return [self._build_marker(f"https://img.example/post-{i}.jpg") for i in range(post_count)]
-            return []
-
-        async def await_timeout(*_:Any, **__:Any) -> None:
-            raise TimeoutError("Image upload timeout exceeded")
-
-        with (
-            pytest.raises(TimeoutError, match = rf"Expected 2, found {expected_found} processed"),
-            self._mock_upload_dependencies(test_bot, file_input, find_all_side_effect, await_timeout),
-        ):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        file_input.send_file.assert_any_await(image_a)
-        file_input.send_file.assert_any_await(image_b)
-
-    @pytest.mark.asyncio
-    async def test_upload_images_succeeds_when_new_markers_exceed_baseline(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-    ) -> None:
-        """Only marker delta beyond baseline should satisfy completion when thumbnails are absent."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        file_input = MagicMock()
-        file_input.send_file = AsyncMock()
-
-        stale_marker = self._build_marker("https://img.example/stale.jpg")
-        marker_a = self._build_marker("https://img.example/a.jpg")
-        marker_b = self._build_marker("https://img.example/b.jpg")
-        marker_query_count = 0
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, **_:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    return [stale_marker]  # baseline before upload
-                return [stale_marker, marker_a, marker_b]  # 2 new markers beyond baseline
-            return []
-
-        async def await_side_effect(condition:Callable[[], Awaitable[bool]], **_:Any) -> bool:
-            if await condition():
-                return True
-            raise TimeoutError("condition did not pass")
-
-        with self._mock_upload_dependencies(test_bot, file_input, find_all_side_effect, await_side_effect):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        file_input.send_file.assert_any_await(image_a)
-        file_input.send_file.assert_any_await(image_b)
-
-    @pytest.mark.asyncio
-    async def test_upload_images_baseline_capture_timeout_defaults_to_zero(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-    ) -> None:
-        """If baseline marker lookup times out, marker fallback should still work with baseline=0."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        file_input = MagicMock()
-        file_input.send_file = AsyncMock()
-
-        marker_a = self._build_marker("https://img.example/a.jpg")
-        marker_b = self._build_marker("https://img.example/b.jpg")
-        marker_query_count = 0
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, **_:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    raise TimeoutError("baseline markers unavailable")
-                return [marker_a, marker_b]
-            return []
-
-        async def await_side_effect(condition:Callable[[], Awaitable[bool]], **_:Any) -> bool:
-            if await condition():
-                return True
-            raise TimeoutError("condition did not pass")
-
-        with self._mock_upload_dependencies(test_bot, file_input, find_all_side_effect, await_side_effect):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        file_input.send_file.assert_any_await(image_a)
-        file_input.send_file.assert_any_await(image_b)
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        ("baseline_count", "post_count"),
-        [
-            pytest.param(0, 2, id = "no_baseline"),
-            pytest.param(1, 3, id = "one_stale_plus_two_new"),
-            pytest.param(2, 4, id = "two_stale_plus_two_new"),
-        ],
-    )
-    async def test_upload_images_marker_delta_determines_completion(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        tmp_path:Path,
-        baseline_count:int,
-        post_count:int,
-    ) -> None:
-        """Completion should succeed when marker delta reaches expected count."""
-        ad_cfg, image_a, image_b = self._build_two_image_ad(base_ad_config, tmp_path)
-
-        file_input = MagicMock()
-        file_input.send_file = AsyncMock()
-
-        marker_query_count = 0
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, **_:Any) -> list[MagicMock]:
-            nonlocal marker_query_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                marker_query_count += 1
-                if marker_query_count == 1:
-                    return [self._build_marker(f"https://img.example/stale-{i}.jpg") for i in range(baseline_count)]
-                return [self._build_marker(f"https://img.example/post-{i}.jpg") for i in range(post_count)]
-            return []
-
-        async def await_side_effect(condition:Callable[[], Awaitable[bool]], **_:Any) -> bool:
-            if await condition():
-                return True
-            raise TimeoutError("condition did not pass")
-
-        with self._mock_upload_dependencies(test_bot, file_input, find_all_side_effect, await_side_effect):
-            await getattr(test_bot, "_upload_images")(ad_cfg)
-
-        file_input.send_file.assert_any_await(image_a)
-        file_input.send_file.assert_any_await(image_b)
-
-
-class TestImageCleanupInPublishAd:
-    """Regression tests for image cleanup loop in publish_ad."""
-
-    @pytest.mark.asyncio
-    async def test_existing_images_removed_before_upload(
-        self,
-        test_bot:KleinanzeigenBot,
-        base_ad_config:dict[str, Any],
-        mock_page:MagicMock,
-        tmp_path:Path,
-    ) -> None:
-        """Cleanup should probe and click remove buttons before upload."""
-        test_bot.page = mock_page
-        test_bot.page.url = "https://www.kleinanzeigen.de/p-anzeige-aufgeben-bestaetigung.html?adId=12345"
-
-        image_path = tmp_path / "img.jpg"
-        ad_cfg = Ad.model_validate(base_ad_config | {"images": [str(image_path)]})
-        ad_cfg = ad_cfg.model_copy(update = {"id": 12345})
-        image_path.write_bytes(b"\xff\xd8\xff")
-        ad_cfg_orig = ad_cfg.model_dump()
-        ad_file = str(tmp_path / "ad.yaml")
-
-        probe_call_count = 0
-        remove_buttons:list[MagicMock] = []
-        event_log:list[str] = []
-
-        async def probe_side_effect(selector_type:By, selector_value:str, **_:Any) -> Element | None:
-            nonlocal probe_call_count
-            if selector_type == By.CSS_SELECTOR and selector_value == "button[aria-label='Bild entfernen']":
-                probe_call_count += 1
-                if probe_call_count <= 3:
-                    remove_btn = MagicMock()
-                    remove_btn.click = AsyncMock(side_effect = lambda idx = probe_call_count: event_log.append(f"remove-{idx}"))
-                    remove_buttons.append(remove_btn)
-                    return remove_btn
-                return None
-            return None
-
-        async def find_all_side_effect(selector_type:By, selector_value:str, *_:Any, **__:Any) -> list[MagicMock]:
-            if selector_type == By.CSS_SELECTOR and selector_value == "input[name^='adImages'][name$='.url']":
-                markers:list[MagicMock] = []
-                for index in range(3):
-                    marker = MagicMock()
-                    marker.attrs.value = f"https://img.example/{index}.jpg"
-                    markers.append(marker)
-                return markers
-            return []
-
-        async def execute_side_effect(script:str) -> Any:
-            if "window.location.href" in script:
-                return test_bot.page.url
-            return None
-
-        async def upload_side_effect(*_:Any, **__:Any) -> None:
-            event_log.append("upload")
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "_dismiss_consent_banner", new_callable = AsyncMock),
-            patch("kleinanzeigen_bot.publishing_form.set_category", new_callable = AsyncMock),
-            patch.object(test_bot, "_set_special_attributes", new_callable = AsyncMock),
-            patch.object(test_bot, "_set_shipping", new_callable = AsyncMock),
-            patch.object(test_bot, "web_input", new_callable = AsyncMock),
-            patch.object(test_bot, "web_probe", new_callable = AsyncMock, side_effect = probe_side_effect),
-            patch.object(test_bot, "web_check", new_callable = AsyncMock, return_value = False),
-            patch.object(test_bot, "web_click", new_callable = AsyncMock),
-            patch.object(test_bot, "web_execute", side_effect = execute_side_effect),
-            patch.object(test_bot, "web_scroll_page_down", new_callable = AsyncMock),
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch("kleinanzeigen_bot.publishing_form.set_contact_fields", new_callable = AsyncMock),
-            patch.object(test_bot, "_upload_images", new_callable = AsyncMock, side_effect = upload_side_effect) as mock_upload,
-            patch("kleinanzeigen_bot.captcha_flow.check_and_wait_for_captcha", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock),
-            patch.object(test_bot, "_web_find_all_once", new_callable = AsyncMock, side_effect = find_all_side_effect),
-            patch.object(test_bot, "web_await", new_callable = AsyncMock, return_value = True),
-        ):
-            await test_bot.publish_ad(ad_file, ad_cfg, ad_cfg_orig, [], AdUpdateStrategy.MODIFY)
-
-        assert sum(button.click.await_count for button in remove_buttons) == 3
-        mock_upload.assert_awaited_once()
-        assert event_log == ["remove-1", "remove-2", "remove-3", "upload"]
